daemon: split frontend service by domain
